+++ HopperRender/HopperRender.h
@@ -54,6 +54,7 @@ class CHopperRender : public CTransformFilter,
 		int* piNeighborScalar,
 		int* piBlackLevel,
 		int* piWhiteLevel,
+		int* piSceneChangeThreshold,
 		int* piIntActiveState,
 		double* pdSourceFPS,
 		double* pdOFCCalcTime,
@@ -63,9 +64,10 @@ class CHopperRender : public CTransformFilter,
 		int* piDimX,
 		int* piDimY,
 		int* piLowDimX,
-		int* piLowDimY) override;
+		int* piLowDimY,
+		unsigned int* piTotalFrameDelta) override;
 	STDMETHODIMP UpdateUserSettings(bool bActivated, int iFrameOutput, double dTargetFPS, bool bUseDisplayFPS,
-							        int iDeltaScalar, int iNeighborScalar, int iBlackLevel, int iWhiteLevel) override;
+							        int iDeltaScalar, int iNeighborScalar, int iBlackLevel, int iWhiteLevel, int iSceneChangeThreshold) override;
 
     // ISpecifyPropertyPages interface
     STDMETHODIMP GetPages(CAUUID* pPages) override;
@@ -120,5 +122,5 @@ class CHopperRender : public CTransformFilter,
 	double m_dBlendingScalar; // Blends from frame 1 to frame 2 (0.0 shows 100% frame 1, 1.0 shows 100% frame 2)
 	bool m_bUseDisplayFPS; // Whether to use the display refresh rate as target FPS
 	bool m_bValidFrameTimes; // Whether valid frame times have been received
-	bool m_bDisableHDR; // Overide which forces SDR (fixes D3D11 H/W decoding causing HDR to be used incorrectly)
+	unsigned int m_iSceneChangeThreshold; // Threshold for scene change detection (total frame delta value)
 };