+++ HopperRender/HopperRenderSettings.h
@@ -27,6 +27,8 @@ private:
 	int m_iNeighborScalar;
 	int m_iBlackLevel;
 	int m_iWhiteLevel;
+	int m_iSceneChangeThreshold;
+	unsigned int m_iTotalFrameDelta;
 	ActiveState m_iIntActiveState; // The state of the filter (0: Deactivated, 1: Not Needed, 2: Active, 3: Too Slow)
 	double m_dSourceFPS; // The source frames per second
 	double m_dTargetFPS;