+++ HopperRender/resource.h
@@ -32,6 +32,7 @@
 #define IDC_USEDISPLAYFPS               1033
 #define IDC_AVGOFCCALCTIME              1034
 #define IDC_PEAKOFCCALCTIME             1035
+#define IDC_TOTALFRAMEDELTA             1036
 
 // Next default values for new objects
 //